Update Qt.py
this is the last update it should have kills from winners and win loss in mongo. youll see two tables that look the same yet one is storing the losers data and the other is the winners data. ill start running this tonight when i get home since I cant right now. 

youll see the L at the end and the W which im sure youll be able to figure out.